Split Postgres storage into repositories
